price-intel-discovery-git-ready/pyproject.toml
[project]
name = "price-intel-discovery"
version = "0.3.0"
description = "Price intelligence prototype: search-engine discovery (API-based) + merchant truth extraction + real total cost."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "pydantic>=2.6",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "httpx>=0.27",
  "selectolax>=0.3.21",
  "python-dotenv>=1.0",
  "APScheduler>=3.10",
  "rapidfuzz>=3.6",
  "rich>=13.7",
]

[tool.ruff]
line-length = 100
